+++ apps/ledger-live-desktop/src/mvvm/components/SideBar/useSideBarViewModel.ts
@@ -207,7 +207,6 @@ export function useSideBarViewModel(): SideBarViewModel {
   }, [location.pathname, push]);
 
   const handleOpenSendModal = useCallback(() => {
-    console.log("called open send flow from sidebar");
     maybeRedirectToAccounts();
     trackEntry("send");
     openSendFlow();
