+++ apps/ledger-live-desktop/src/mvvm/features/Send/hooks/useOpenSendFlow.ts
@@ -10,14 +10,13 @@ import type { EnhancedModularDrawerConfiguration } from "@ledgerhq/live-common/w
 import { closeDialog, openDialog } from "~/renderer/reducers/modularDrawer";
 import { HOOKS_TRACKING_LOCATIONS } from "~/renderer/analytics/hooks/variables";
 import { setOriginFlow } from "~/renderer/analytics/originFlow";
-import { getLLDCoinFamily } from "~/renderer/families";
 
 const SEND_ACCOUNT_SELECTION_DRAWER_CONFIGURATION: EnhancedModularDrawerConfiguration = {
   assets: { rightElement: "balance" },
   networks: {},
 };
 
-export type WorkflowParams = {
+type WorkflowParams = {
   account?: AccountLike;
   parentAccount?: Account;
   recipient?: string;
@@ -36,7 +35,7 @@ export function useOpenSendFlow() {
     (params?: WorkflowParams) => {
       setOriginFlow(HOOKS_TRACKING_LOCATIONS.sendModal);
 
-      const defaultOpenSendFlowImpl = (nextParams?: WorkflowParams) => {
+      const openSendFlowImpl = (nextParams?: WorkflowParams) => {
         if (!nextParams?.account) {
           // When there are no accounts, the old modal requires an account and would throw.
           // Always use the new drawer for account selection (or empty state) in that case.
@@ -49,7 +48,7 @@ export function useOpenSendFlow() {
                 dialogConfiguration: SEND_ACCOUNT_SELECTION_DRAWER_CONFIGURATION,
                 onAccountSelected: (account: AccountLike, parentAccount?: Account) => {
                   dispatch(closeDialog());
-                  defaultOpenSendFlowImpl({
+                  openSendFlowImpl({
                     ...nextParams,
                     account,
                     parentAccount,
@@ -76,14 +75,6 @@ export function useOpenSendFlow() {
 
         const family = getFamilyFromAccount(nextParams.account, nextParams.parentAccount ?? null);
         const shouldUseNewFlow = isEnabledForFamily(family);
-        const specific = family ? getLLDCoinFamily(family) : null;
-
-        console.log("use open send flow", family, specific?.openSendFlow, shouldUseNewFlow);
-
-        if (specific?.openSendFlow) {
-          specific.openSendFlow(dispatch, nextParams);
-          return;
-        }
 
         if (shouldUseNewFlow) {
           let normalizedAmount: string | undefined;
@@ -118,7 +109,7 @@ export function useOpenSendFlow() {
         }
       };
 
-      defaultOpenSendFlowImpl(params);
+      openSendFlowImpl(params);
     },
     [hasNoAccounts, dispatch, isEnabledForFamily, getFamilyFromAccount],
   );
