+++ apps/ledger-live-desktop/src/renderer/families/aleo/constants.ts
@@ -1,6 +1,5 @@
 export const PRIVATE_BALANCE_PLACEHOLDER = "***";
 
 export enum AleoCustomModal {
-  SEND = "MODAL_ALEO_SEND",
   SELF_TRANSFER = "MODAL_ALEO_SELF_TRANSFER",
 }
