+++ apps/ledger-live-desktop/src/renderer/families/aleo/index.ts
@@ -1,18 +1,18 @@
 import AccountBalanceSummaryFooter from "./AccountBalanceSummaryFooter";
 import ModularDrawerAddAccountFlowManager from "./ModularDrawerAddAccountFlowManager";
 import accountHeaderManageActions from "./AccountHeaderManageActions";
+import createSendSteps from "./createSendSteps";
 import StepSummaryAmountRow from "./modals/send/steps/StepSummaryAmountRow";
 import StepSummaryNetworkFeesRow from "./modals/send/steps/StepSummaryNetworkFeesRow";
 import StepSummaryPostAlert from "./modals/send/steps/StepSummaryPostAlert";
 import operationDetails from "./operationDetails";
-import openSendFlow from "./openSendFlow";
 import type { AleoFamily } from "./types";
 
 const family: AleoFamily = {
   AccountBalanceSummaryFooter,
   ModularDrawerAddAccountFlowManager,
   accountHeaderManageActions,
-  openSendFlow,
+  createSendSteps,
   operationDetails,
   StepSummaryAmountRow,
   StepSummaryNetworkFeesRow,
