+++ apps/ledger-live-desktop/src/renderer/families/aleo/modals/send/steps/StepRecipient.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { isSelfTransferTransaction } from "@ledgerhq/live-common/families/aleo/utils";
-import { SendStepRecipient } from "~/renderer/families/aleo/SendModal/SendStepRecipient";
+import { SendStepRecipient } from "~/renderer/families/aleo/SendTransferModal/AleoSendStepRecipient";
 import { SelfTransferStepRecipient } from "~/renderer/families/aleo/SelfTransferModal/SelfTransferStepRecipient";
 import type { StepProps } from "~/renderer/modals/Send/types";
